keep track of template being processed
